\name{AQDMRS.list}
\alias{AQDMRS.list}
\title{AQDMRS.list}
\usage{
  AQDMRS.list(...)
}
\arguments{
  \item{\dots}{query arguments}
}
\description{
  Metadata interface to the AQDMRS gateway.
}
\references{
  \url{http://www.epa.gov/airdata/tas_URL_Query_Construction_Details_list.html}
}
\seealso{
  Other AQDMRS: \code{\link{AQDMRS.data}},
  \code{\link{AQDMRS.query}}
}

