% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_list.R
\name{file_list}
\alias{file_list}
\title{File list

Get a tibble listing all of the .parquet files on luftdaten}
\usage{
file_list()
}
\value{
A tibble containing a list of files and metadata.
}
\description{
File list

Get a tibble listing all of the .parquet files on luftdaten
}
\examples{
file_list()
}
