% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_list.R
\name{get_file_list}
\alias{get_file_list}
\title{Get File List}
\usage{
get_file_list()
}
\value{
A list of .parquet files
}
\description{
This function accesses the luftdaten website (http://archive.luftdaten.info/parquet/) and downloads a
list of available parquet files containing project data.
}
