% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_file_list.R
\name{html_file_list}
\alias{html_file_list}
\title{HTML file list

Get a HTML file table from luftdaten and turn it into a tibble}
\usage{
html_file_list(url)
}
\arguments{
\item{url}{A luftdaten site url.}
}
\value{
A tibble containing a list of files and metadata.
}
\description{
HTML file list

Get a HTML file table from luftdaten and turn it into a tibble
}
\examples{
html_file_list("http://archive.luftdaten.info/parquet/")
}
