% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_download_file.R
\name{safe_download_file}
\alias{safe_download_file}
\title{Safe Download File

Download a file from a url (using curl), if there is an error, return this as a data.frame.}
\usage{
safe_download_file(url, name, destination_folder)
}
\arguments{
\item{url}{url of file.}

\item{name}{destination file name.}

\item{destination_folder}{destination folder name.}
}
\value{
A data.frame containing details of the downloaded file, or error detials.
}
\description{
Safe Download File

Download a file from a url (using curl), if there is an error, return this as a data.frame.
}
